improve chirp create and edit forms, add telegram-like uploading animation, fix a lot of bugs

+++ app/Http/Requests/StoreLikeRequest.php
@@ -14,7 +14,7 @@ class StoreLikeRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'chirp_id' => ['required', 'numeric']
+            'chirp_id' => ['required', 'integer']
         ];
     }
 }
